multimodal-sentiment/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
deepface>=0.0.79
streamlit>=1.28.0
plotly>=5.17.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
tf-keras>=2.15.0
fer
